refactor(ai): extract shared exact dice-odds table into diceOdds.js
The d6-convolution win-probability table is needed by every expected-value
AI. Move it out of ai_claude into src/ai/diceOdds.js as a single source of
truth (the incoming ai_codex strategy reuses it instead of copying it).

- diceOdds.js: WIN_TABLE + winProbability (ties to defender, dice clamped to
  8, out-of-range returns 0).
- ai_claude.js: import the shared table; re-export winProbability for its
  existing importers (tests, tooling).

No behavior change: table values are identical to the prior copy; the
ai_claude win-probability tests still pass.
